+++ src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { ButtonSizeEnum, ButtonVariantBean, CardVariants, ComboItemBean, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, StatusTagStatus, TableHeaderSize, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean } from "./beans";
+import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, CardVariants, ComboItemBean, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, StatusTagStatus, TableHeaderSize, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean } from "./beans";
 import { LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
 import { ListItemBean } from "./beans/index.js";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
@@ -51,6 +51,28 @@ export namespace Components {
           * theme variant, default 'dark'
          */
         "theme"?: ThemeVariantBean;
+    }
+    interface ZAvatar {
+        /**
+          * [optional] Avatar background color
+         */
+        "backgroundColor"?: string;
+        /**
+          * [optional] Avatar image
+         */
+        "image"?: string;
+        /**
+          * [optional] Avatar size
+         */
+        "size"?: AvatarSize;
+        /**
+          * [optional] Avatar text
+         */
+        "text"?: string;
+        /**
+          * [optional] Avatar text color
+         */
+        "textColor"?: string;
     }
     interface ZBody {
         "component": string;
@@ -834,7 +856,9 @@ export namespace Components {
     }
     interface ZNavigationTab {
         "disabled"?: boolean;
+        "orientation": string;
         "selected"?: boolean;
+        "size": string;
     }
     interface ZNavigationTabs {
         /**
@@ -845,6 +869,28 @@ export namespace Components {
           * Available sizes: `big` and `small`. Defaults to `big`.
          */
         "size"?: TabSizeBean;
+    }
+    interface ZNotification {
+        /**
+          * action button text
+         */
+        "actiontext"?: string;
+        /**
+          * icon on the left of the content
+         */
+        "contenticonname"?: string;
+        /**
+          * enable close icon
+         */
+        "showclose"?: boolean;
+        /**
+          * enable shadow
+         */
+        "showshadow"?: boolean;
+        /**
+          * alert variant type
+         */
+        "type": NotificationType;
     }
     interface ZPaginationBar {
         /**
@@ -1191,6 +1237,12 @@ declare global {
         prototype: HTMLZAppTopbarElement;
         new (): HTMLZAppTopbarElement;
     };
+    interface HTMLZAvatarElement extends Components.ZAvatar, HTMLStencilElement {
+    }
+    var HTMLZAvatarElement: {
+        prototype: HTMLZAvatarElement;
+        new (): HTMLZAvatarElement;
+    };
     interface HTMLZBodyElement extends Components.ZBody, HTMLStencilElement {
     }
     var HTMLZBodyElement: {
@@ -1455,6 +1507,12 @@ declare global {
         prototype: HTMLZNavigationTabsElement;
         new (): HTMLZNavigationTabsElement;
     };
+    interface HTMLZNotificationElement extends Components.ZNotification, HTMLStencilElement {
+    }
+    var HTMLZNotificationElement: {
+        prototype: HTMLZNotificationElement;
+        new (): HTMLZNotificationElement;
+    };
     interface HTMLZPaginationBarElement extends Components.ZPaginationBar, HTMLStencilElement {
     }
     var HTMLZPaginationBarElement: {
@@ -1591,6 +1649,7 @@ declare global {
         "z-app-header": HTMLZAppHeaderElement;
         "z-app-switcher": HTMLZAppSwitcherElement;
         "z-app-topbar": HTMLZAppTopbarElement;
+        "z-avatar": HTMLZAvatarElement;
         "z-body": HTMLZBodyElement;
         "z-button": HTMLZButtonElement;
         "z-button-filter": HTMLZButtonFilterElement;
@@ -1635,6 +1694,7 @@ declare global {
         "z-myz-list-item": HTMLZMyzListItemElement;
         "z-navigation-tab": HTMLZNavigationTabElement;
         "z-navigation-tabs": HTMLZNavigationTabsElement;
+        "z-notification": HTMLZNotificationElement;
         "z-pagination-bar": HTMLZPaginationBarElement;
         "z-pagination-page": HTMLZPaginationPageElement;
         "z-panel-elem": HTMLZPanelElemElement;
@@ -1705,6 +1765,28 @@ declare namespace LocalJSX {
           * theme variant, default 'dark'
          */
         "theme"?: ThemeVariantBean;
+    }
+    interface ZAvatar {
+        /**
+          * [optional] Avatar background color
+         */
+        "backgroundColor"?: string;
+        /**
+          * [optional] Avatar image
+         */
+        "image"?: string;
+        /**
+          * [optional] Avatar size
+         */
+        "size"?: AvatarSize;
+        /**
+          * [optional] Avatar text
+         */
+        "text"?: string;
+        /**
+          * [optional] Avatar text color
+         */
+        "textColor"?: string;
     }
     interface ZBody {
         "component"?: string;
@@ -2376,7 +2458,11 @@ declare namespace LocalJSX {
          */
         "modaltitle"?: string;
         /**
-          * emitted on close icon click, returns modalid
+          * emitted on background click, returns modalid
+         */
+        "onModalBackgroundClick"?: (event: CustomEvent<any>) => void;
+        /**
+          * emitted on close button click, returns modalid
          */
         "onModalClose"?: (event: CustomEvent<any>) => void;
         /**
@@ -2573,7 +2659,9 @@ declare namespace LocalJSX {
     interface ZNavigationTab {
         "disabled"?: boolean;
         "onSelected"?: (event: CustomEvent<any>) => void;
+        "orientation"?: string;
         "selected"?: boolean;
+        "size"?: string;
     }
     interface ZNavigationTabs {
         /**
@@ -2584,6 +2672,36 @@ declare namespace LocalJSX {
           * Available sizes: `big` and `small`. Defaults to `big`.
          */
         "size"?: TabSizeBean;
+    }
+    interface ZNotification {
+        /**
+          * action button text
+         */
+        "actiontext"?: string;
+        /**
+          * icon on the left of the content
+         */
+        "contenticonname"?: string;
+        /**
+          * notification action event
+         */
+        "onNotificationAction"?: (event: CustomEvent<any>) => void;
+        /**
+          * notification close event
+         */
+        "onNotificationClose"?: (event: CustomEvent<any>) => void;
+        /**
+          * enable close icon
+         */
+        "showclose"?: boolean;
+        /**
+          * enable shadow
+         */
+        "showshadow"?: boolean;
+        /**
+          * alert variant type
+         */
+        "type"?: NotificationType;
     }
     interface ZPaginationBar {
         /**
@@ -2935,6 +3053,7 @@ declare namespace LocalJSX {
         "z-app-header": ZAppHeader;
         "z-app-switcher": ZAppSwitcher;
         "z-app-topbar": ZAppTopbar;
+        "z-avatar": ZAvatar;
         "z-body": ZBody;
         "z-button": ZButton;
         "z-button-filter": ZButtonFilter;
@@ -2979,6 +3098,7 @@ declare namespace LocalJSX {
         "z-myz-list-item": ZMyzListItem;
         "z-navigation-tab": ZNavigationTab;
         "z-navigation-tabs": ZNavigationTabs;
+        "z-notification": ZNotification;
         "z-pagination-bar": ZPaginationBar;
         "z-pagination-page": ZPaginationPage;
         "z-panel-elem": ZPanelElem;
@@ -3010,6 +3130,7 @@ declare module "@stencil/core" {
             "z-app-header": LocalJSX.ZAppHeader & JSXBase.HTMLAttributes<HTMLZAppHeaderElement>;
             "z-app-switcher": LocalJSX.ZAppSwitcher & JSXBase.HTMLAttributes<HTMLZAppSwitcherElement>;
             "z-app-topbar": LocalJSX.ZAppTopbar & JSXBase.HTMLAttributes<HTMLZAppTopbarElement>;
+            "z-avatar": LocalJSX.ZAvatar & JSXBase.HTMLAttributes<HTMLZAvatarElement>;
             "z-body": LocalJSX.ZBody & JSXBase.HTMLAttributes<HTMLZBodyElement>;
             "z-button": LocalJSX.ZButton & JSXBase.HTMLAttributes<HTMLZButtonElement>;
             "z-button-filter": LocalJSX.ZButtonFilter & JSXBase.HTMLAttributes<HTMLZButtonFilterElement>;
@@ -3054,6 +3175,7 @@ declare module "@stencil/core" {
             "z-myz-list-item": LocalJSX.ZMyzListItem & JSXBase.HTMLAttributes<HTMLZMyzListItemElement>;
             "z-navigation-tab": LocalJSX.ZNavigationTab & JSXBase.HTMLAttributes<HTMLZNavigationTabElement>;
             "z-navigation-tabs": LocalJSX.ZNavigationTabs & JSXBase.HTMLAttributes<HTMLZNavigationTabsElement>;
+            "z-notification": LocalJSX.ZNotification & JSXBase.HTMLAttributes<HTMLZNotificationElement>;
             "z-pagination-bar": LocalJSX.ZPaginationBar & JSXBase.HTMLAttributes<HTMLZPaginationBarElement>;
             "z-pagination-page": LocalJSX.ZPaginationPage & JSXBase.HTMLAttributes<HTMLZPaginationPageElement>;
             "z-panel-elem": LocalJSX.ZPanelElem & JSXBase.HTMLAttributes<HTMLZPanelElemElement>;