+++ src/components/navigation/z-navigation-tabs/index.spec.ts
@@ -8,16 +8,17 @@ describe("Suite test ZNavigationTabs", () => {
     const page = await newSpecPage({
       components: [ZNavigationTabs],
       html: `
-      <z-navigation-tabs role="tablist" orientation="horizontal" size="big">
-        <z-navigation-tab size="big">
-          <z-icon name="house" slot="icon"></z-icon>
-          <span slot="label">Home</span>
-        </z-navigation-tab>
-      </z-navigation-tabs>
-    `});
+        <z-navigation-tabs role="tablist" orientation="horizontal" size="big">
+          <z-navigation-tab size="big">
+            <z-icon name="house" slot="icon"></z-icon>
+            <span slot="label">Home</span>
+          </z-navigation-tab>
+        </z-navigation-tabs>
+      `
+    });
 
     expect(page.root).toEqualHtml(`
-      <z-navigation-tabs role="tablist" orientation="horizontal" size="big">
+      <z-navigation-tabs class="interactive-1" role="tablist" orientation="horizontal" size="big">
         <mock:shadow-root>
           <nav>
             <slot></slot>
@@ -35,16 +36,17 @@ describe("Suite test ZNavigationTabs", () => {
     const page = await newSpecPage({
       components: [ZNavigationTabs],
       html: `
-      <z-navigation-tabs role="tablist" orientation="horizontal" size="small">
-        <z-navigation-tab size="small">
-          <z-icon name="house" slot="icon"></z-icon>
-          <span slot="label">Home</span>
-        </z-navigation-tab>
-      </z-navigation-tabs>
-    `});
+        <z-navigation-tabs role="tablist" orientation="horizontal" size="small">
+          <z-navigation-tab size="small">
+            <z-icon name="house" slot="icon"></z-icon>
+            <span slot="label">Home</span>
+          </z-navigation-tab>
+        </z-navigation-tabs>
+      `
+    });
 
     expect(page.root).toEqualHtml(`
-      <z-navigation-tabs role="tablist" orientation="horizontal" size="small">
+      <z-navigation-tabs class="interactive-2" role="tablist" orientation="horizontal" size="small">
         <mock:shadow-root>
           <nav>
             <slot></slot>
@@ -62,16 +64,17 @@ describe("Suite test ZNavigationTabs", () => {
     const page = await newSpecPage({
       components: [ZNavigationTabs],
       html: `
-      <z-navigation-tabs role="tablist" size="big" orientation="vertical">
-        <z-navigation-tab size="big">
-          <z-icon name="house" slot="icon"></z-icon>
-          <span slot="label">Home</span>
-        </z-navigation-tab>
-      </z-navigation-tabs>
-    `});
+        <z-navigation-tabs role="tablist" size="big" orientation="vertical">
+          <z-navigation-tab size="big">
+            <z-icon name="house" slot="icon"></z-icon>
+            <span slot="label">Home</span>
+          </z-navigation-tab>
+        </z-navigation-tabs>
+      `
+    });
 
     expect(page.root).toEqualHtml(`
-      <z-navigation-tabs role="tablist" size="big" orientation="vertical">
+      <z-navigation-tabs class="interactive-1" role="tablist" size="big" orientation="vertical">
         <mock:shadow-root>
           <nav>
             <slot></slot>
@@ -84,5 +87,4 @@ describe("Suite test ZNavigationTabs", () => {
       </z-navigation-tabs>
     `);
   });
-
 });